# packages/contracts/src/generated/wolfari/identity/v1/identity.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.12.4
//   protoc               unknown
// source: wolfari/identity/v1/identity.proto

/* eslint-disable */
import type { Metadata } from "@grpc/grpc-js";
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";
import { Timestamp } from "../../../google/protobuf/timestamp";
import { AccountStatus, EmailPurpose, SystemRole } from "../../common/v1/types";

export const protobufPackage = "wolfari.identity.v1";

export interface ValidateSessionRequest {
  access_token?: string | undefined;
}

export interface ValidateSessionResponse {
  user_id?: string | undefined;
  system_role?: SystemRole | undefined;
  status?: AccountStatus | undefined;
  expires_at?: Timestamp | undefined;
}

export interface GetProfilesRequest {
  user_ids?: string[] | undefined;
}

export interface Profile {
  id?:
    | string
    | undefined;
  /** Projection of Identity users.full_name. */
  display_name?: string | undefined;
  avatar?: string | undefined;
}

export interface GetProfilesResponse {
  profiles?: Profile[] | undefined;
}

export interface GetAccountEmailDeliveryRequest {
  token_id?: string | undefined;
}

export interface AccountEmailDelivery {
  user_id?: string | undefined;
  recipient_email?: string | undefined;
  purpose?:
    | EmailPurpose
    | undefined;
  /** Sensitive RPC-only field: never publish in events. */
  one_time_link?: string | undefined;
  expires_at?: Timestamp | undefined;
}

export interface GetAccountEmailDeliveryResponse {
  valid?:
    | boolean
    | undefined;
  /** Present iff valid. */
  delivery?: AccountEmailDelivery | undefined;
}

export const WOLFARI_IDENTITY_V1_PACKAGE_NAME = "wolfari.identity.v1";

/** Source: DDL/API/Event v1.0 section 6. Authorization is enforced by the owner. */

export interface IdentityServiceClient {
  validateSession(request: ValidateSessionRequest, metadata?: Metadata): Observable<ValidateSessionResponse>;

  getProfiles(request: GetProfilesRequest, metadata?: Metadata): Observable<GetProfilesResponse>;

  getAccountEmailDelivery(
    request: GetAccountEmailDeliveryRequest,
    metadata?: Metadata,
  ): Observable<GetAccountEmailDeliveryResponse>;
}

/** Source: DDL/API/Event v1.0 section 6. Authorization is enforced by the owner. */

export interface IdentityServiceController {
  validateSession(
    request: ValidateSessionRequest,
    metadata?: Metadata,
  ): Promise<ValidateSessionResponse> | Observable<ValidateSessionResponse> | ValidateSessionResponse;

  getProfiles(
    request: GetProfilesRequest,
    metadata?: Metadata,
  ): Promise<GetProfilesResponse> | Observable<GetProfilesResponse> | GetProfilesResponse;

  getAccountEmailDelivery(
    request: GetAccountEmailDeliveryRequest,
    metadata?: Metadata,
  ):
    | Promise<GetAccountEmailDeliveryResponse>
    | Observable<GetAccountEmailDeliveryResponse>
    | GetAccountEmailDeliveryResponse;
}

export function IdentityServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["validateSession", "getProfiles", "getAccountEmailDelivery"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("IdentityService", method)(constructor.prototype[method], method, descriptor);
      Object.defineProperty(constructor.prototype, method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("IdentityService", method)(constructor.prototype[method], method, descriptor);
      Object.defineProperty(constructor.prototype, method, descriptor);
    }
  };
}

export const IDENTITY_SERVICE_NAME = "IdentityService";